milestone 3
So for the implementation of the toggle for light and dark mode, I made a button that has an image that switches between a moon icon for dark mode and a sun icon for light mode. The script element is in the head, so I put DOMContentLoaded, so that the javascript logic will only begin after the entire document object model loads. So first part of the js code was basically making sure that by default the users page looks a certain way, as their browser must have set, and the second part (on click onwards) is for everytime the button is pressed. Specifically, when the button is pressed or toggled, we get the attribute that contained theme in the html tag document.documentElement) to check what it is, if it is dark we change it to light, and vice versa. Everytime we also reassign the local storage so that after refreshes, the users preferences stay intact.