milestone 8
Added a bit of file config allowing user to only upload png, jpg, jpeg and gifs, though most of this including uploading and ensuring users upload secure files is left.

Made a few changes in profile page but most of the displaying is left

primarily worked on the details template and backend logic

worked on the get part, post part is left

In the get part:

With the help of user_id (email) i got the id, and then used the id (sessionid variable) to query details from the other tables

if there is no data present in the database, some default values are set (black and white pic for profile pic, rest are just blank in order to show placeholders)
If data is present on the other hand, it will be displayed in the details page (edit profile details) so the user can just make changes to their current description or details rather than writing all over again. This works for all points, just have to check for the profile pic part because i have to save user pics in the instances folder which is private