milestone 10
scrapped supabase for now, will complete phase 1 or the cs50 version of this project with all files on my system including the database being mysql, later for production may move this all to amazon rds or something that can host my database, or use backend as a service in something like supabase.

Changes ive made:
in the details route, made a system in the post request, where the user uploads a file and clicks save. The file is then checked for validity, and then I split the email into first and last name, like i did in the profile page.
Then I made a filename with an f string to join the first and last name and the file type. This joined with user-pics ahead of it will be the database path, which is the path stored in the database, which is a url part ahead of the normal routes. This is a relative path to the file stored in the servers files.

----important---

so basically when the user clicks on details or profile html, a lot of get request are sent for each picture, which is done normally by html.
The src attribute in an HTML <img> tag essentially acts as the endpoint for a GET request. In this case when a get request is sent on this route: '/user-pics/<filename>' it invokes the flask route in the backend, which in turn serves the picture

Now after this, I constructed an upload folder which is the absolute path in the device, not in the database. Make a directory by joining the app.instance_path or basically in the instance folder make a user-pics folder if it doesnt exists already, and within that make another path for the file in the device.