some changes
made a menu that has options where the user can choose if they want to search by name, bio, skills, interests or softskills. Only some front end of this menu bar is done, with some javascript that makes sure that the menu is opened when the user clicks on the filter button. The JavaScript in the first function ensures that when the button is clicked, the menu is toggled open or closed by adding or removing the show class, depending on whether it was already present. The second part listens for any click on the document and checks if the clicked element is outside both the filterButton and filterDropdown. If it is, the show class is removed to close the menu.