milestone 14
Finished the javascript part of the search page completely. So basically made sure that if the filterValue, or the option the user selected is "name" then javascript searches by the name header, otherwise it searches by the description. The text in the elements is compared to the text the user inputted, and if the element's text has contains what the user has inputted, it will be displayed with the default css styles, else the display would not be shown.
This is wrapped in a function, which is triggered every time the user enters any input, or changes the option to something else (the last two event listeners are about this).