milestone 16
So this line:
{% if option == selected_option %}selected{% endif %}>{{ option }}
ensures that the choice remains throughout reloads. This ensures that whenever the choice is changed and the form is reloaded, the filterValue in javascript doesnt get "" (due to no selection), but the instead, it gets the previous selection: 'Name', enabling the search to work properly for the name search. How this actually works is:
When the user selects an option, and it is equal to the selected option
(How it is checked if they are equal:
if the user selects "Name" and submits the form, the URL becomes /search?filter=Name, and request.args.get('filter') captures "Name", in the backend. selected_option is then passed to the search.html template in render_template. Inside the template, {% if option == selected_option %}selected{% endif %} checks each option against selected_option and adds the selected attribute to the dropdown option that matches.)
, then jinja inserts selected in the html, making it the default selected option throughout reloads, ensuring that the choice isnt lost after the form is submitted and a reload happens.