milestone 17
Made a visiting page. It takes the users id who you are visiting, from the url. This happens after the user clicks on the name or the profile picture in the any of the search page's profile previews. This is because the search page sends the user to a url with the id parameter of the user whose preview you are seeing (via <a href="/visiting/{{ user.id }}">). Then with almost the same route and templating as the profile page, the profile of person who you are visiting will be displayed via the visiting template.