milestone 20
made a followers feature, where a user can follow and unfollow users, this is stored in a followers table, have to check if i need following table. The count is stored in the users table, check idk this is necessary too. After that did responsiveness of profile page and visiting page, refining needed tho